Relocated Header.js & Home.js;
