+++ src/App.css
@@ -37,10 +37,6 @@
   }
 }
 
-.editableBoard {
-  top: 90px;
-}
-
 .testAbility {
   color: red;
 }
