+++ src/index.js
@@ -5,7 +5,7 @@ import App from './App';
 import Test from './react-web/Comp/test';
 import reportWebVitals from './reportWebVitals';
 import Editor from './react-web/Editor';
-import Home from './react-web/Home';
+// import Home from './nav/Home';
 import Login from './login/Login';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
@@ -35,8 +35,8 @@ ReactDOM.render(
             <Routes>
                 <Route path='/' element={<App />}>
                     {/* <Route path='/' element={<Main />}> */}
-                    <Route index element={<Home />} />
-                    <Route path='login' element={<Login />} />
+                    {/* <Route index element={<Home />} /> */}
+                    <Route index element={<Login />} />
                     <Route path='editor' element={<Editor />} />
                     <Route path='test' element={<Test />} />
                     <Route path="*" element={
